Refactor scripts and remove unused scripts
