+++ UI/src/components/Swap.tsx
@@ -6,7 +6,7 @@ import tokenList from '../constants/tokenList.json'
 import { Token } from '../constants/Interfaces'
 import { notification } from './notifications'
 import { useDebouncedEffect } from '../utils/useDebouncedEffect'
-import {findQuote, swap} from '../sdk/quoter/front_communication'
+import {findQuote, swap} from '../sdk/swap/front_communication'
 import {Quote} from '../sdk/types'
 import './Swap.scss'
 import RoutingDiagram from './RoutingDiagram'
