+++ UI/src/contracts/addresses.ts
@@ -1,8 +1,8 @@
 
 // chainId: 42161    
-export const RATE_X_ADDRESS = "0xEBc7C8b072b413719451d1a6Ae3ab576e4006D9d"
-export const UNISWAP_HELPER_ADDRESS = "0xb69ecb00DA1De1545f67b08285d35E798370C995"
-export const SUSHISWAP_HELPER_ADDRESS = "0xf75753b6ff857c0B5848786522f3f636d33494ca"
-export const CURVE_HELPER_ADDRESS = "0x66A4EB7858e42cDe046456a09e5bC43EBC39127b"
-export const CAMELOT_HELPER_ADDRESS = "0xd34377236040337b414419D7D5Bf3AE96321C893"
-export const BALANCER_HELPER_ADDRESS = "0xD498C1eCa61C10666F10003D6683a22560B3C360"
+export const RATE_X_ADDRESS = "0xBe15c474FD09c1f7F563093E83a51Ee195C74435"
+export const UNISWAP_HELPER_ADDRESS = "0x7dA3EcbCc2033BD7F0FA52a53B1e4948b661FFB2"
+export const SUSHISWAP_HELPER_ADDRESS = "0xe79Cdd48110ADA9C9c516984852aaCB5EAf63f28"
+export const CURVE_HELPER_ADDRESS = "0x29E79B27Ed2b1d44e7Aff8F75BE1b91aDbfFF3EC"
+export const CAMELOT_HELPER_ADDRESS = "0xA800Cc484C4b8F5cd0f7758B6B0df9688754d9d2"
+export const BALANCER_HELPER_ADDRESS = "0xB2b9211B24903c5935dA8eBDEBFF79f877d0B7bf"
