+++ UI/src/sdk/swap/solidity_communication.ts
@@ -3,7 +3,6 @@ import { fetchPoolsData } from './graph_communication'
 import { ERC20_ABI } from '../../contracts/abi/common/ERC20_ABI'
 import initRPCProvider from '../../providers/RPCProvider'
 import Web3 from 'web3'
-import {TQuoteUniLike} from "../routing/uni_like_algo/types";
 import { RateXContract } from '../../contracts/rateX/RateX'
 import {findRoute} from "../routing/main";
 
@@ -41,7 +40,7 @@ async function executeSwap(
     // @ts-ignore
     await tokenInContract.methods.approve(RateXContract.options.address, amountIn).send({ from: signer })
     let transactionHash: string = ''
-    quote = transformQuoteForSolidity(quote)
+    quote = transferQuoteWithBalancerPoolIdToAddress(quote)
 
     console.log("usao u swap");
 
@@ -58,13 +57,12 @@ async function executeSwap(
   }
 }
 
-/* Function to transform the quote to be compatible with the solidity contract
-  * The solidity contract expects the poolId to be an address, but the graph returns it as a bytes32
-  * The solidity contract also expects the token names to be removed
+/* Function to transform poolId from bytes32 to address
+  * The solidity contract expects the poolId to be an address, but the graph for balancer returns it as a bytes32
   * @param quote: The quote to be transformed
   * @returns The transformed quote
 */
-function transformQuoteForSolidity(quote: Quote): Quote {
+function transferQuoteWithBalancerPoolIdToAddress(quote: Quote): Quote {
   quote.routes[0].swaps.map((swap) => {
     if (swap.poolId.length === 66) {
       swap.poolId = swap.poolId.slice(0, 42)      // convert to address
