--- contracts/scripts/uniswap/exprimental/README.md
@@ -1,3 +1,0 @@
-This folder has uni offchain quoter and other utils functions from sdk copied for easier testing
-It is exprimental and it can be helpful later for debugging as it is easier to test various contracts with hardhat
-If we wanted to test it in sdk, we would need to deploy all contracts as we don't have hardhat environment there
